Restructure and refactor shared content
