--- Frontend/src/features/shared/constants/index.ts
@@ -1,1 +1,0 @@
-export const BASE_URL = 'http://localhost:5166/api';
