--- Frontend/src/features/shared/utilities/index.ts
@@ -1,2 +1,0 @@
-export * from './fetchJson';
-export * from './fetchWithToken';
