+++ Frontend/src/router/router.tsx
@@ -1,6 +1,6 @@
 import { createBrowserRouter } from 'react-router';
 import Layout from '../pages/Layout';
-import { Login } from '../components/Login';
+import Login from '../components/Login';
 import { requireAuthLoader } from '../utilities/loaders/requireAuthLoader';
 import { Companies, Company } from '../features/companies/components';
 import { companiesLoader, companyLoader } from '../features/companies/loaders';
