+++ Frontend/src/utilities/api/fetchJson.ts
@@ -1,4 +1,4 @@
-import { CustomError } from '../classes';
+import { CustomError } from '../../utilities/classes';
 
 export async function fetchJson<T>(input: RequestInfo | URL, init?: RequestInit): Promise<T> {
   const res = await fetch(input, init);
