+++ Frontend/src/utilities/classes.ts
