+++ Frontend/src/utilities/constants.ts
@@ -1,1 +1,2 @@
 export const TOKENS = 'tokens';
+export const BASE_URL = 'http://localhost:5166/api';
